requirements.txt
pyyaml
py-trello
psycopg2-binary
python-dotenv
sqlalchemy